+++ module/data/actor/common.js
@@ -93,18 +93,18 @@ export class CommonActorData extends foundry.abstract.TypeDataModel {
 
   /**
    *
-   * @param {object} data the data object to migrate
+   * @param {object} source the data object to migrate
    */
-  static migrateData(data) {
-    if (data.other?.cosm !== undefined) data.other.cosm = migrateCosm(data.other.cosm);
+  static migrateData(source) {
+    if (source.other?.cosm !== undefined) source.other.cosm = migrateCosm(source.other.cosm);
 
-    for (const attribute of Object.keys(data.attributes ?? {})) {
-      if (typeof data?.attributes?.[attribute] === 'number') {
-        data.attributes[attribute] = { base: data.attributes[attribute] };
+    for (const attribute of Object.keys(source.attributes ?? {})) {
+      if (typeof source?.attributes?.[attribute] === 'number') {
+        source.attributes[attribute] = { base: source.attributes[attribute] };
       }
     }
 
-    for (const skill of Object.values(data.skills ?? {})) {
+    for (const skill of Object.values(source.skills ?? {})) {
       if (Object.hasOwn(skill, 'adds') && typeof skill.adds !== 'number') {
         let skillAdd = parseInt(skill.adds);
         skillAdd = isNaN(skillAdd) ? 0 : skillAdd;
@@ -115,10 +115,10 @@ export class CommonActorData extends foundry.abstract.TypeDataModel {
       }
     }
 
-    if (Object.hasOwn(data, "other.possibilities") && typeof data.other.possibilities === 'number') {
-      data.other.possibilities = { value: data.other.possibilities }
+    if (Object.hasOwn(source?.other, "possibilities") && typeof source.other.possibilities === 'number') {
+      source.other.possibilities = { value: source.other.possibilities }
     }
-    return super.migrateData(data);
+    return super.migrateData(source);
   }
 
   /**